pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monopoly-family-edition-desktop"
version = "1.0.0"
description = "Python desktop launcher for Monopoly Family Edition"
requires-python = ">=3.10"
dependencies = [
  "pywebview==5.3.2"
]

[project.scripts]
monopoly-desktop = "python_app.app:main"
